Conflicts:
	src/main/resources/build.number
	src/main/resources/version.number

+++ src/main/modules/admin/R/smip.R
@@ -22,7 +22,13 @@
 #smart.install.packages(url.package="http://www.math.utu.fi/projects/software/bio/ROTS_1.1.1.tar.gz")
 #smart.install.scavenge.web.packages("http://brainarray.mbni.med.umich.edu/Brainarray/Database/CustomCDF/16.0.0/entrezg.asp")
 
-smart.install.scavenge.web.packages <- function(url, chiptype="all") {
+smart.install.scavenge.web.packages <- function(url, chiptype="all", update=0, list.only=0) {
+	# Name of the static pages
+	if(length(grep("latest", url)) > 0) {
+		web.page <- paste(readLines(url), collapse="")
+		latest.build <- gsub("\'.*$", "", gsub("^.*URL=\\.\\.", "", web.page))
+		url <- gsub("\\/latest\\/.*$", latest.build, url);
+	}
 
 	# Parse URL
 	url.root <- substring(url, 1, gregexpr("/[^/]*$", url)[[1]][1])
@@ -43,18 +49,23 @@ smart.install.scavenge.web.packages <- function(url, chiptype="all") {
 		links <- links[grep(chiptype, links)];
 	}
 
+	# if list.only == 1, only the list of available packages will returned
+	if(list.only == 1) {
+		return(links)
+	}
+
 	# Install each linked package
 	for (link in links) {
-		smart.install.packages(url.package=link)
+		smart.install.packages(url.package=link, update=update)
 	}
 }
 
-smart.install.bioconductor.repo <- function(repo.index, mirror=NA) {
+smart.install.bioconductor.repo <- function(repo.index, mirror=NA, update=0) {
 	
 	# check what version of Bionconductor is being used
 	bioc.intaller.loaded <- try(library(BiocInstaller))
 	if(class(bioc.intaller.loaded) == "try-error") {
-		smart.install.packages(bioconductor.package="BiocInstaller", mirror=mirror)
+		smart.install.packages(bioconductor.package="BiocInstaller", mirror=mirror, update=update)
 	}
 	library(BiocInstaller)
 	current.bioc.version <- biocVersion()
@@ -62,14 +73,14 @@ smart.install.bioconductor.repo <- function(repo.index, mirror=NA) {
 	# Install available annotation packages
 	current.bioc.url <- paste("http://www.bioconductor.org/packages/", current.bioc.version, "/data/annotation", sep="");
 	for (package in available.packages(contrib.url(current.bioc.url))[,"Package"]) {
-		smart.install.packages(bioconductor.package = package, mirror = mirror)
+		smart.install.packages(bioconductor.package = package, mirror = mirror, update=update)
 	}
 
 	# Select the given repository
 	#orig.repos <- setRepositories(ind=c(repo.index))
 	#
 	#for (package in available.packages()[,"Package"]) {
-	#	smart.install.packages(bioconductor.package = package, mirror = mirror)
+	#	smart.install.packages(bioconductor.package = package, mirror = mirror, update=update)
 	#}
 	#
 	# Restore original repositories
@@ -85,7 +96,7 @@ smart.install.bioconductor.repo <- function(repo.index, mirror=NA) {
 	#}
 }
 
-smart.install.packages <- function(package=NA, bioconductor.package=NA, url.package=NA, mirror=NA) {
+smart.install.packages <- function(package=NA, bioconductor.package=NA, url.package=NA, mirror=NA, update=0) {
 	
 	# Check parameters
 	package.defs <- c(package, bioconductor.package, url.package)
@@ -100,13 +111,15 @@ smart.install.packages <- function(package=NA, bioconductor.package=NA, url.pack
 		package.name <- gsub(".*/(.*)_.*", "\\1", url.package)
 	}
 
-	if (!is.installed(package.name)) {
-		cat(paste("Will now install", package.name, "\n"))
-	} else {
-		cat(paste("Already installed", package.name, "\n"))
-		return(invisible(TRUE))
+	if(update==0) {
+		if (!is.installed(package.name)) {
+			cat(paste("Will now install", package.name, "\n"))
+		} else {
+			cat(paste("Already installed", package.name, "\n"))
+			return(invisible(TRUE))
+		}
 	}
-	
+
 	if (!is.na(package)) {
 		
 		repos = ifelse(is.na(mirror), getOption("repos"), mirror)
@@ -124,10 +137,13 @@ smart.install.packages <- function(package=NA, bioconductor.package=NA, url.pack
 		
 		# Download URL to temp file, install and remove
 		tempfile.path <- tempfile("package", fileext=".tar.gz")
-		download.file(url=url.package, destfile=tempfile.path)
-		install.packages(pkgs=c(tempfile.path), repos=NULL)
-		unlink(tempfile.path)
-		
+		a <- try(download.file(url=url.package, destfile=tempfile.path))
+		if(class(a) != "try-error") {
+			install.packages(pkgs=c(tempfile.path), repos=NULL)
+			unlink(tempfile.path)
+		} else {
+			warning(paste("package", url.package, "is not valid a web-page", sep=" "))
+		}
 		
 	} else {
 		stop("Must specify something to install");
@@ -148,3 +164,70 @@ is.installed <- function(package) {
 		return(package %in% rownames(installed.packages()))
 	}
 }
+
+check.affy.customnames <- function(script.basename, db.custom.packages) {
+
+	# Find R scripts used to normalize affy chips
+	r.script.basename <- gsub("/admin/", "/microarray/", script.basename);
+	all.packages <- rownames(installed.packages())
+	all.r.scripts <- list.files(path=r.script.basename);
+	affy.r.script <- all.r.scripts[grep("norm-affy", all.r.scripts)];
+
+	# Install stringdist package. Needed for finding partial matches
+	smart.install.packages(package="stringdist", mirror=repo.cran)
+	library("stringdist");
+
+	# Polish up custom-CDF names
+	if(length(db.custom.packages) > 0) {
+		db.custom.packages <- gsub(".*/(.*)_.*", "\\1", db.custom.packages)
+	} else {
+		stop("Must specify a list of custom-CDF packages");
+	}
+
+	# For each affy-norm script, find code defining which custom-CDF packages are supported
+	supported.custom.packages <- NULL;
+	for(i in 1:length(affy.r.script)) {
+		r.script <- scan(file=file.path(r.script.basename, affy.r.script[i]), what="", sep="\n")
+
+		#Find instances in SADL descriptions
+		if(length(grep("PARAMETER custom\\.chiptype", r.script) > 0)) {
+			sadl.row <- r.script[grep("PARAMETER custom\\.chiptype", r.script)];
+			sadl.row <- gsub("^.*\\[", "", sadl.row);
+			sadl.row <- gsub("\\].*$", "", sadl.row);
+			packages.in.sadl.row <- unlist(strsplit(sadl.row, "\\s*,\\s*"))
+			for(j in 1:length(packages.in.sadl.row)) {
+				custom.package <- unlist(strsplit(packages.in.sadl.row[j], "\\s*:\\s*"))[2];
+				custom.package <- gsub("\\(.+\\)", "cdf", custom.package);
+				supported.custom.packages <- c(supported.custom.packages, custom.package);
+			}
+		}
+
+		#Find other instances where parameter custom_cdf has been used
+		if(length(grep("custom_cdf\\s*<-|custom_cdf\\s*=", r.script) > 0)) {
+			rscript.row <- r.script[grep("custom_cdf\\s*<-|custom_cdf\\s*=", r.script)];
+			rscript.row <- gsub("^.*<-|^.*=", "", rscript.row);
+			rscript.row <- gsub("\\s+|\"", "", rscript.row);
+			supported.custom.packages <- c(supported.custom.packages, rscript.row);
+		}
+	}
+
+	# Check if the package exists
+	for(j in 1:length(supported.custom.packages)) {
+		if(!(supported.custom.packages[j] %in% db.custom.packages)) {
+			cat(paste("Package", supported.custom.packages[j], "in", affy.r.script[i], "not found\n"));
+			partial.match <- db.custom.packages[ain(db.custom.packages, supported.custom.packages[j], maxDist=3)];
+			if(length(partial.match) > 0) {
+				for(k in 1:length(partial.match)) {
+					if(partial.match[k] %in% rownames(installed.packages()) == TRUE) {
+						cat(paste("\tConsider using", partial.match[k], "\n"));
+					} else {
+						cat(paste("\tConsider installing and using", partial.match[k], "\n"));
+					}
+				}
+			} else {
+				cat(paste("\tPackage", supported.custom.packages[j], "in", affy.r.script[i], "has not matches in current custom-CDF database\n"));
+			}
+		}
+	}
+}
+
